コードのリファクタリングとNgmoduleの分割
