+++ src/.priettierrc
@@ -1,0 +1,9 @@
+
+	{
+		"printWidth": 100,
+		"trailingComma": "es5",
+		"tabWidth": 2,
+		"semi": true,
+		"singleQuote": true,
+		"endOfLine": "lf"
+	}
