--- src/app/app.component.css
@@ -1,120 +1,0 @@
-.page {
-    overflow: scroll;
-    max-width: 960px;
-    height: 100%;
-    margin: 0 auto;
-    padding: 10px 0;
-    border-left: 1px solid #DDDDDD;
-    border-right: 1px solid #DDDDDD;
-}
-
-
-section {
-    margin: 50px 10px 90px;
-}
-
-.chat-editor {
-    display: block;
-    position: fixed;
-    bottom: 0;
-    width: 100%;
-    max-width: 958px;
-    min-height: 100px;
-    margin: 0;
-    padding: 10px;
-    border-top: 1px solid #8a7575;
-    background-color: #fff;
-}
-
-.page .chat .media-body {
-    padding: 15px;
-    margin: 0 10px;
-    background-color: #f5f5f5;
-    border-radius: 4px;
-    font-size: 15px;
-    word-break: break-all;
-}
-
-.media-body h4.media-heading {
-    color: #aaaaaa;
-    font-size: 13px;
-}
-
-.media-body .media-date {
-    color: #aaaaaa;
-    font-size: 11px;
-}
-
-.media-body pre {
-    font-family: inherit;
-    color: inherit;
-    font-size: inherit;
-    margin: 0;
-    word-wrap: break-word;
-    white-space: pre-wrap;
-}
-
-.icon {
-    display: block;
-    width: 50px;
-    height: 50px;
-    margin: 5px;
-    border-radius: .3rem;
-    color: #fff;
-    font-size: 30px;
-    line-height: 50px;
-    text-align: center;
-}
-
-.edit-form {
-    width: 100%;
-}
-
-.edit-btn {
-    margin: 0 5px;
-}
-
-.media-left .icon {
-    background-color: #f0974e;
-}
-.media-right .icon {
-    background-color: #35b8c2;
-}
-
-.media-body h4.media-heading .fa {
-    cursor: pointer;
-}
-
-.media-body h4.media-heading .fa:hover {
-    color: #666;
-}
-
-.media-body div.edit-buttons {
-    margin-top: 10px;
-    text-align: center;
-}
-
-.media-body div.edit-buttons .btn {
-    width: 100px;
-}
-
-.media-body h4.media-heading .fa {
-    cursor: pointer;
-}
-    
-.media-body h4.media-heading .fa:hover {
-    color: #666;
-}
-
-.media-body div.edit-buttons {
-    margin-top: 10px;
-    text-align: center;
-}
-
-.media-body div.edit-buttons .btn {
-    width: 100px;
-}
-
-.sky-blue {
-    background: #7cbde8;
-}
