+++ src/app/app.module.ts
@@ -1,23 +1,14 @@
 import { NgModule } from '@angular/core';
-import { FormsModule } from '@angular/forms';
 import { BrowserModule } from '@angular/platform-browser';
 
 import { AppComponent } from './app.component';
-import { CommentDatePipe } from './pipes/comment-date.pipe';
-import { initializeApp, provideFirebaseApp } from '@angular/fire/app';
-import { environment } from '../environments/environment';
-import { provideAuth, getAuth } from '@angular/fire/auth';
-import { provideFirestore, getFirestore } from '@angular/fire/firestore';
+import { ChatComponent } from './chat/chat.component';
+import { CoreModule } from './core/core.module';
+import { SharedModule } from './shared/shared.module';
 
 @NgModule({
-  declarations: [AppComponent, CommentDatePipe],
-  imports: [
-    BrowserModule,
-    FormsModule,
-    provideFirebaseApp(() => initializeApp(environment.firebase)),
-    provideAuth(() => getAuth()),
-    provideFirestore(() => getFirestore()),
-  ],
+  declarations: [AppComponent, ChatComponent],
+  imports: [BrowserModule, SharedModule, CoreModule],
   providers: [],
   bootstrap: [AppComponent],
 })
