+++ src/app/chat/chat.component.html
@@ -1,0 +1,92 @@
+<div class="page sky-blue">
+  <section class="chat">
+    <div class="chat-timeline">
+      <div
+        *ngFor="let comment of comments$ | async"
+        class="list-group-item list-group-item-action flex-column align-items-start border-0 sky-blue"
+      >
+        <div class="media">
+          <div *ngIf="comment.user.status === false" class="media-left">
+            <a href="#" class="icon">{{ comment.user.initial }}</a>
+          </div>
+          <div class="media-body">
+            <div class="d-flex w-100 justify-content-between">
+              <h4 class="media-heading">
+                {{ comment.user.name }}
+                <ng-container *ngIf="comment.user.status === true">
+                  <div>
+                    <i
+                      class="fa fa-edit ml-2"
+                      title="編集"
+                      (click)="comment.isEdit = !comment.isEdit"
+                    ></i>
+                    <i
+                      class="fa fa-trash ml-2"
+                      title="削除"
+                      (click)="deleteComment(comment)"
+                    ></i>
+                  </div>
+                </ng-container>
+              </h4>
+              <small class="media-date">{{
+                comment.date | date: "dd/MM/yyyy"
+              }}</small>
+            </div>
+            <ng-container *ngIf="!comment.isEdit">
+              <pre>{{ comment.message }}</pre>
+            </ng-container>
+            <ng-container *ngIf="comment.isEdit">
+              <div class="input-group">
+                <textarea
+                  [(ngModel)]="changed_comment"
+                  name="edit-comment"
+                  class="edit-form form-cotrol"
+                ></textarea>
+              </div>
+              <div class="edit-buttons">
+                <button
+                  class="edit-btn btn btn-success btn-sm"
+                  (click)="editComment(comment)"
+                >
+                  更新
+                </button>
+                <button class="edit-btn btn btn-warning btn-sm">
+                  キャンセル
+                </button>
+              </div>
+            </ng-container>
+          </div>
+          <div *ngIf="comment.user.status === true" class="media-right">
+            <a href="#" class="icon">{{ comment.user.initial }}</a>
+          </div>
+        </div>
+      </div>
+    </div>
+  </section>
+
+  <section class="chat-editor">
+    <form class="chat-form">
+      <div class="form-header d-flex flex-row-reverse mb-2">
+        <span class="input-group-btn">
+          <button
+            (click)="addComment(comment); comment = ''"
+            class="btn btn-info btn-sm"
+            type="submit"
+            tabindex="2"
+          >
+            送信
+          </button>
+        </span>
+      </div>
+      <div class="input-group">
+        <textarea
+          [(ngModel)]="comment"
+          name="comment"
+          class="form-control"
+          placeholder="メッセージを入力"
+          tabindex="1"
+        ></textarea>
+      </div>
+    </form>
+  </section>
+</div>
