+++ src/app/pipes/comment-date.pipe.ts
@@ -2,13 +2,11 @@ import { formatDate } from '@angular/common';
 import { Pipe, PipeTransform } from '@angular/core';
 
 @Pipe({
-  name: 'commentDate'
+  name: 'commentDate',
 })
 export class CommentDatePipe implements PipeTransform {
-
   transform(value: number, ...args: string[]): string {
-    const format = 'yyyy年MM月dd日 HH:mm'
+    const format = 'yyyy年MM月dd日 HH:mm';
     return formatDate(value, format, 'en-US');
   }
-
 }
